MOBI-290 Medicast upgrade add tablet view add view switcher add edit mode for photo deletion change to single url pattern for mobile/tablet handling
git-svn-id: http://anonsvn.icesoft.org/repo/icemobile/trunk/icemobile@31011 1de00972-ac04-4911-a46b-06793c8d78d9

+++ samples/mvc-mediacast/src/main/java/org/icemobile/samples/spring/mediacast/Media.java
@@ -109,8 +109,10 @@ public class Media implements Serializable {
 	}
 	
 	public void dispose(){
-		file.delete();
-		file = null;
+		if( file != null ){
+			file.delete();
+			file = null;
+		}
 		type = null;
 		fileName = null;
 	}
